jobs/requirements.txt
fastapi
uvicorn
psycopg2-binary
scikit-learn
pandas
numpy
joblib
python-dotenv
matplotlib
